+++ spec/helpers/to_have_outline.js
@@ -1,0 +1,14 @@
+beforeEach(function() {
+  jasmine.addMatchers({
+    toHaveOutline: function(util, customEqualityTesters) {
+      return {
+        compare: function(element) {
+          element = up.element.get(element)
+          return {
+            pass: up.element.styleNumber(element, 'outlineWidth') > 0
+          }
+        }
+      }
+    }
+  })
+})